refactor: quick structural change
