+++ src/manki/data/translation/collins/parse.py
@@ -1,0 +1,30 @@
+import json
+from typing import Dict, List, Optional
+
+from bs4 import BeautifulSoup
+from pydantic import BaseModel
+
+
+class CollinsWordEntry(BaseModel):
+    word: str
+    part_of_speech: Optional[List[str]] = None
+    pronunciation: Optional[str] = None
+    definitions: Optional[List[str]] = None
+    translations: Optional[List[str]] = None
+    examples: Optional[List[str]] = None
+    synonyms: Optional[List[str]] = None
+    antonyms: Optional[List[str]] = None
+    metadata: Optional[Dict[str, str]] = None
+
+
+class CollinsDictionaryParser:
+    def __init__(self, file_path: str):
+        self.file_path = file_path
+
+    def parse(self) -> List[CollinsWordEntry]:
+        with open(self.file_path, "r", encoding="utf-8") as file:
+            html_content = file.read()
+
+        # Implement the actual parsing logic for Collins dictionary here
+        # For now, return an empty list
+        return []
